requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
appnope==0.1.4
asttokens==3.0.1
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
executing==2.2.1
fastapi==0.136.0
filelock==3.28.0
fonttools==4.62.1
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.2
idna==3.11
ipykernel==7.2.0
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.5.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.4
opencv-python==4.13.0.92
packaging==26.1
parso==0.8.6
pexpect==4.9.0
pi_heif==1.3.0
pillow==12.2.0
platformdirs==4.9.6
polars==1.39.3
polars-runtime-32==1.39.3
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.13.3
pydantic_core==2.46.3
Pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.26
PyYAML==6.0.3
pyzmq==27.1.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.5.4
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ultralytics==8.4.37
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.44.0
wcwidth==0.6.0
